Fix invalid Web IDL
The JSON-LD Framing API specification extends interfaces of the JSON-LD API but
does not use the `partial` extension keyword, meaning that interfaces from both
specs cannot be automatically merged into one.

The JSON-LD Framing API specification also re-defines interfaces that are
already defined in the JSON-LD API for no apparent reason, which is bad practice
as it creates two normative definitions of the same object.

This update makes the following changes to drop duplicate definitions and
improve cross-reference links between the JSON-LD Framing API and the JSON-LD
API:
- Dropped duplicate definition of `JsonLd` interface
- Defined `JsonLdProcessor` as `partial` (and dropped `constructor` member since
it is already defined in the JSON-LD API)
- Dropped duplicate definitions of `JsonLdRecord` and `JsonLdInput`
- Defined `JsonLdOptions` as `partial` (and dropped `ordered` since it is
already defined in the JSON-LD API)
- Replaced `data-cite` links with `{{xx}}` links where possible
- Dropped self-reference from `xref` ReSpec config array (useless, local
definitions are always available)